fim do exercicio 1

+++ 4_semestre/Fiap.Exercicio1/Fiap.Exercicio1.UI/Model/Conta.cs
@@ -6,16 +6,19 @@ using System.Threading.Tasks;
 
 namespace Fiap.Banco.Model
 {
+    //Classe abstrata:
+    // 1 - Nao pode ser instanciada
+    // 2 - Pode conter metodos abstratos (sem implementacao)
     public abstract class Conta
     {
 
         public int Agencia { get; set; }
         public int Numero { get; set; }
         public DateTime DataAbertura { get; set; }
-        public double Saldo { get; set; }
+        public decimal Saldo { get; set; }
 
-        public abstract void Depositar(double valor);
-        public abstract void Retirar(double valor);
+        public abstract void Depositar(decimal valor);
+        public abstract void Retirar(decimal valor);
 
     }
